+++ .gitignore
@@ -4,3 +4,6 @@
 tesla-app
 tfetch
 todo.md
+CLAUDE.md
+.claude
+.claudeignore
