+++ cmd/server/main.go
@@ -1,0 +1,5 @@
+package main
+
+func main() {
+	// TODO: Fleet API server implementation
+}
