+++ cmd/tfetch/main.go
@@ -5,13 +5,10 @@ import (
 	"log"
 	"os"
 	"path/filepath"
-	apitypes "tfetch/api/types"
-	"tfetch/api/vehicle-status"
-	"tfetch/api/wake"
 	"tfetch/auth"
-	vehiclecommand "tfetch/command"
-	drawstatus "tfetch/draw-status"
-	services "tfetch/services/get-data-handler"
+	"tfetch/command"
+	"tfetch/model"
+	"tfetch/tesla"
 	"tfetch/ui"
 
 	"github.com/joho/godotenv"
@@ -21,14 +18,14 @@ var ErrorTooManyArgs = errors.New("too many args provided")
 var ErrorInvalidArgs = errors.New("invalid arguments")
 
 type AppDependencies struct {
-	StatusLogger      ui.StatusLoggerMethods
+	StatusLogger      model.StatusLoggerMethods
 	Flag              string
 	AuthService       auth.AuthMethods
-	VehicleService    apitypes.VehicleMethods
-	WakeService       apitypes.WakeMethods
-	DrawStatusService drawstatus.DrawMethods
-	IssueCommand      func(status ui.StatusLoggerMethods, token auth.Token, command string, vehicleService apitypes.VehicleMethods, wakeService apitypes.WakeMethods) error
-	GetDataHandler    func(status ui.StatusLoggerMethods, authService auth.AuthMethods, vehicleService apitypes.VehicleMethods, wakeService apitypes.WakeMethods) (*apitypes.VehicleData, error)
+	VehicleService    model.VehicleMethods
+	WakeService       model.WakeMethods
+	DrawStatusService model.DrawMethods
+	IssueCommand      func(status model.StatusLoggerMethods, token model.Token, command string, vehicleService model.VehicleMethods, wakeService model.WakeMethods) error
+	GetDataHandler    func(status model.StatusLoggerMethods, authService auth.AuthMethods, vehicleService model.VehicleMethods, wakeService model.WakeMethods) (*model.VehicleData, error)
 }
 
 func main() {
@@ -44,7 +41,7 @@ func main() {
 		log.Fatalf("\rUsage: %s", err)
 	}
 
-	var statusLogger ui.StatusLoggerMethods
+	var statusLogger model.StatusLoggerMethods
 	if flag == "-w" {
 		statusLogger = ui.NewNoopLogger()
 	} else {
@@ -58,41 +55,23 @@ func main() {
 		StatusLogger:      statusLogger,
 		Flag:              flag,
 		AuthService:       &auth.AuthService{},
-		VehicleService:    &vehicle.VehicleService{},
-		DrawStatusService: &drawstatus.DrawService{},
-		WakeService:       &wake.WakeService{},
-		IssueCommand:      vehiclecommand.IssueCommand,
-		GetDataHandler:    services.GetDataHandler,
+		VehicleService:    &tesla.VehicleService{},
+		DrawStatusService: &ui.DrawService{},
+		WakeService:       &tesla.WakeService{},
+		IssueCommand:      command.IssueCommand,
+		GetDataHandler:    tesla.GetDataHandler,
 	}
 
 	err = runApp(app)
 	if err != nil {
 		log.Fatalf("\r%s\n", err)
 	}
-	return
 }
 
 func runApp(app AppDependencies) error {
 	vehicleService := app.VehicleService
 	authService := app.AuthService
 	wakeService := app.WakeService
-	// token, err := app.AuthService.CheckLogin(app.StatusLogger)
-	// if err != nil {
-	// 	return err
-	// }
-
-	// no issue command logic right now focus on refactor
-	// if len(args) == 2 {
-	// 	app.Status <- ui.ProgressUpdate{Message: "Issuing command"}
-	// 	command := args[1]
-	//
-	// 	err = app.IssueCommand(app.Status, *token, command, vehicleService)
-	// 	if err != nil {
-	// 		return err
-	// 	}
-	//
-	// 	fmt.Printf("\rCommand %s issued successfully\n", command)
-	// }
 
 	vehicleData, err := app.GetDataHandler(app.StatusLogger, authService, vehicleService, wakeService)
 	if err != nil {
@@ -130,7 +109,6 @@ func loadEnv() error {
 }
 
 func validateFlags(args []string) (string, error) {
-	// one flag allowed for now
 	validFlags := map[string]bool{
 		"-w": true,
 	}
